bell_detector/requirements.txt
pyAudioAnalysis==0.3.5
sounddevice==0.4.0
python-pushover==0.4
SoundFile==0.10.3.post1
matplotlib==3.1.2
simplejson==3.16.0
scipy==1.4.1
numpy==1.18.1
hmmlearn==0.2.2
eyeD3==0.8.12
pydub==0.23.1
scikit_learn==0.21.3
tqdm==4.44.1
plotly==4.1.1
